rename old histogram file to legacy, will be removed later
