+++ templatefitter/__init__.py
@@ -1,4 +1,4 @@
-from templatefitter.histogram import *
+from templatefitter.histogram_legacy import *
 from templatefitter.nll import *
 from templatefitter.fitter import *
 from templatefitter.minimizer import *
