+++ templatefitter/histogram_legacy.py
